# fix summator performance - change wrappers to primitive type (int), r…