--- a/hw04-gc/build.gradle
+++ b/hw04-gc/build.gradle
@@ -1,0 +1,6 @@
+plugins {
+    id 'java'
+}
+
+sourceCompatibility = JavaVersion.VERSION_17
+targetCompatibility = JavaVersion.VERSION_17