--- a/hw04-gc/src/main/java/homework/src/main/java/ru/calculator/CalcDemo.java
+++ b/hw04-gc/src/main/java/homework/src/main/java/ru/calculator/CalcDemo.java
@@ -1,0 +1,39 @@
+package homework.src.main.java.ru.calculator;
+
+
+/*
+-Xms256m
+-Xmx256m
+-XX:+HeapDumpOnOutOfMemoryError
+-XX:HeapDumpPath=./logs/heapdump.hprof
+-XX:+UseG1GC
+-Xlog:gc=debug:file=./logs/gc-%p-%t.log:tags,uptime,time,level:filecount=5,filesize=10m
+*/
+
+
+import java.time.LocalDateTime;
+
+public class CalcDemo {
+    public static void main(String[] args) {
+        long counter = 100_000_000;
+        var summator = new Summator();
+        long startTime = System.currentTimeMillis();
+
+        for (var idx = 0; idx < counter; idx++) {
+            var data = new Data(idx);
+            summator.calc(data);
+
+            if (idx % 10_000_000 == 0) {
+                System.out.println(LocalDateTime.now() + " current idx:" + idx);
+            }
+        }
+
+        long delta = System.currentTimeMillis() - startTime;
+        System.out.println(summator.getPrevValue());
+        System.out.println(summator.getPrevPrevValue());
+        System.out.println(summator.getSumLastThreeValues());
+        System.out.println(summator.getSomeValue());
+        System.out.println(summator.getSum());
+        System.out.println("spend msec:" + delta + ", sec:" + (delta / 1000));
+    }
+}