--- a/hw04-gc/src/main/java/homework/src/main/java/ru/calculator/Data.java
+++ b/hw04-gc/src/main/java/homework/src/main/java/ru/calculator/Data.java
@@ -1,0 +1,13 @@
+package homework.src.main.java.ru.calculator;
+
+public class Data {
+    private final Integer value;
+
+    public Data(Integer value) {
+        this.value = value;
+    }
+
+    public Integer getValue() {
+        return value;
+    }
+}