--- a/hw04-gc/src/main/java/homework/src/main/java/ru/calculator/Summator.java
+++ b/hw04-gc/src/main/java/homework/src/main/java/ru/calculator/Summator.java
@@ -1,0 +1,49 @@
+package homework.src.main.java.ru.calculator;
+
+public class Summator {
+    private int sum = 0;
+    private int prevValue = 0;
+    private int prevPrevValue = 0;
+    private int sumLastThreeValues = 0;
+    private int someValue = 0;
+    private int size = 0;
+
+    //!!! сигнатуру метода менять нельзя
+    public void calc(Data data) {
+        size += 1;
+        if (size % 6_600_000 == 0) {
+            size = 0;
+        }
+        sum += data.getValue();
+
+        sumLastThreeValues = data.getValue() + prevValue + prevPrevValue;
+
+        prevPrevValue = prevValue;
+        prevValue = data.getValue();
+
+        for (var idx = 0; idx < 3; idx++) {
+            someValue += (sumLastThreeValues * sumLastThreeValues / (data.getValue() + 1) - sum);
+            someValue = Math.abs(someValue) + size;
+        }
+    }
+
+    public Integer getSum() {
+        return sum;
+    }
+
+    public Integer getPrevValue() {
+        return prevValue;
+    }
+
+    public Integer getPrevPrevValue() {
+        return prevPrevValue;
+    }
+
+    public Integer getSumLastThreeValues() {
+        return sumLastThreeValues;
+    }
+
+    public Integer getSomeValue() {
+        return someValue;
+    }
+}