… hw04-gc/README.md
… settings.gradle