fix summator performance - change wrappers to primitive type (int), replace List<Data> to int variable.

1. В первой части задания ощутимый прирост производительности прекратился после значений (**-Xms1280m -Xmx1280m**), а сравнялся после (**-Xms1536m -Xmx1536m**). Данные запусков:
```
-Xms256m -Xmx256m -XX:+UseG1GC
spend msec:13503, sec:13

-Xms2048m -Xmx2048m -XX:+UseG1GC
spend msec:7413, sec:7

-Xms1024m -Xmx1024m -XX:+UseG1GC
spend msec:7718, sec:7

-Xms512m -Xmx512m -XX:+UseG1GC
spend msec:8509, sec:8

-Xms768m -Xmx768m -XX:+UseG1GC
spend msec:8152, sec:8

-Xms896m -Xmx896m -XX:+UseG1GC
spend msec:7844, sec:7

-Xms1280m -Xmx1280m -XX:+UseG1GC
spend msec:7531, sec:7

-Xms1408m -Xmx1408m -XX:+UseG1GC
spend msec:7667, sec:7

-Xms1536m -Xmx1536m -XX:+UseG1GC
spend msec:7499, sec:7
```
2. После оптимизаций в классе **Summator** на при любых настройках выделяемой памяти, значение оставалось одинаково низким - 
`spend msec:441, sec:0`